# openapi_actions.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "World Engine GPT Actions API",
    "description": "Persistent authoritative world/game-state API for the World Engine GPT, including the World Engine Turn Protocol (WETP-1.0), capability routing, a universal entity/relationship graph, knowledge provenance, a bounded context compiler, deterministic tabletop-RPG rules, WORLD/SCENE spatial layers, off-screen simulation, automatic image cues, persistent visual continuity, and a shadow-safe narrative director/dialogue/prose-quality compiler. ChatGPT normalizes intent and renders prose; this service owns facts, routing, context selection, mutations, narrative contracts, and audit records.",
    "version": "4.3.0"
  },
  "paths": {
    "/api/campaign": {
      "post": {
        "summary": "Ensure Campaign",
        "description": "Create a campaign if absent and return its authoritative clock, weather and revision.",
        "operationId": "ensureCampaign",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ensurecampaign",
                  "properties": {}
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-openai-isConsequential": false
      }
    },
    "/api/setup/character": {
      "post": {
        "summary": "Save Character",
        "description": "Create/update a character. Intended for campaign setup or explicit authoritative edits.",
        "operationId": "saveCharacter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Savecharacter",
                  "properties": {}
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-openai-isConsequential": false
      }
    },
    "/api/setup/location": {
      "post": {
        "summary": "Save Location",
        "description": "Create/update persistent location description, tags and structured state.",
        "operationId": "saveLocation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Savelocation",
                  "properties": {}
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-openai-isConsequential": false
      }
    },
...
    "/api/turn": {
      "post": {
        "summary": "Resolve Turn",
        "description": "Route one normalized player turn through the World Engine Turn Protocol (WETP-1.0).\n\nChatGPT supplies bounded intent objects; the backend selects capabilities,\ncompiles relevant context, executes deterministic providers, records\nidempotency/revisions, and returns an auditable result packet.",
        "operationId": "resolveTurn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Resolveturn",
                  "properties": {}
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-openai-isConsequential": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CampaignRequest": {
        "properties": {
          "campaign_id": {
            "type": "string",
            "title": "Campaign Id",
            "default": "default"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": "World Engine Campaign"
          },
          "world_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "World Time"
          }
        },
        "type": "object",
        "title": "CampaignRequest"
      },
      "CharacterRequest": {
        "properties": {
          "campaign_id": {
            "type": "string",
            "title": "Campaign Id",
            "default": "default"
          },
          "character_id": {
            "type": "string",
            "title": "Character Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "level": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Level",
            "default": 1
          },
          "hp": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Hp",
            "default": 1
          },
          "max_hp": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Hp",
            "default": 1
          },
          "ac": {
            "type": "integer",
            "maximum": 40.0,
            "minimum": 1.0,
            "title": "Ac",
            "default": 10
          },
          "location": {
            "type": "string",
            "title": "Location",
            "default": "unknown"
          },
          "abilities": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Abilities",
            "description": "Ability modifiers, e.g. {str:3,dex:2}",
            "properties": {}
          },
          "proficiency_bonus": {
            "type": "integer",
            "title": "Proficiency Bonus",
            "default": 2
          },
          "conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Conditions"
          },
          "resources": {
            "additionalProperties": true,
            "type": "object",
            "title": "Resources",
            "properties": {}
          },
          "inventory": {
            "items": {},
            "type": "array",
            "title": "Inventory"
          },
          "notes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Notes",
            "properties": {}
          }
        },
        "type": "object",
        "required": [
          "character_id",
          "name"
        ],
        "title": "CharacterRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocationRequest": {
        "properties": {
          "campaign_id": {
            "type": "string",
            "title": "Campaign Id",
            "default": "default"
          },
          "location_id": {
            "type": "string",
            "title": "Location Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "region": {
            "type": "string",
            "title": "Region",
            "default": "unknown"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          },
          "realm_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realm Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "state": {
            "additionalProperties": true,
            "type": "object",
            "title": "State",
            "properties": {}
          }
        },
        "type": "object",
        "required": [
          "location_id",
          "name"
        ],
        "title": "LocationRequest"
      },
...
      "ResolveTurnRequest": {
        "properties": {
          "campaign_id": {
            "type": "string",
            "title": "Campaign Id",
            "default": "default"
          },
          "actor_kind": {
            "anyOf": [
              {
                "type": "string",
                "const": "character"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Kind"
          },
          "actor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "player_text": {
            "type": "string",
            "maxLength": 20000,
            "title": "Player Text",
            "default": ""
          },
          "intents": {
            "items": {
              "$ref": "#/components/schemas/TurnIntentRequest"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Intents"
          },
          "expected_revision": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "mode": {
            "type": "string",
            "enum": [
              "execute",
              "plan",
              "context_only",
              "capabilities"
            ],
            "title": "Mode",
            "default": "execute"
          },
          "max_context_chars": {
            "type": "integer",
            "maximum": 60000.0,
            "minimum": 2000.0,
            "title": "Max Context Chars",
            "default": 14000
          },
          "include_archive": {
            "type": "boolean",
            "title": "Include Archive",
            "default": false
          },
          "continue_on_error": {
            "type": "boolean",
            "title": "Continue On Error",
            "default": false
          },
          "retry_failed": {
            "type": "boolean",
            "title": "Retry Failed",
            "default": false
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "image_trigger_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "scene_start",
                  "battle_start",
                  "new_location",
                  "event_choice",
                  "character_reference",
                  "npc_reference"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Trigger Type"
          },
          "image_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Summary"
          },
          "choice_options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Choice Options"
          },
          "decision_phase": {
            "type": "string",
            "enum": [
              "before",
              "after"
            ],
            "title": "Decision Phase",
            "default": "before"
          },
          "major_consequence": {
            "type": "boolean",
            "title": "Major Consequence",
            "default": false
          },
          "narrative_mode_override": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "off",
                  "shadow",
                  "compare",
                  "enforce"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative Mode Override"
          },
          "narrative_hint": {
            "additionalProperties": true,
            "type": "object",
            "title": "Narrative Hint",
            "description": "Optional presentation-only inputs such as scene_function, speaker_id, speech_act, facts_to_reveal, information_to_withhold, subtext, and mechanically_supported_player_effects. These cannot overwrite authoritative state.",
            "properties": {}
          }
        },
        "type": "object",
        "title": "ResolveTurnRequest"
      },
...
      "TurnIntentRequest": {
        "properties": {
          "intent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent Id",
            "description": "Stable identifier within this turn; defaults to intent_1, intent_2, etc."
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type",
            "description": "Friendly intent alias such as move, check, attack, interact, relation, fact, or advance_time."
          },
          "capability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capability",
            "description": "Explicit capability ID; takes precedence over type when supplied."
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters",
            "properties": {}
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Depends On"
          },
          "optional": {
            "type": "boolean",
            "title": "Optional",
            "default": false
          }
        },
        "type": "object",
        "title": "TurnIntentRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context",
            "properties": {}
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "https://YOUR-HOST"
    }
  ]
}